+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### 🚀 Performance
 
 1. 添加bridge invoke性能检测逻辑(开发模式下默认开启).
+2. (ContentScript)减少职位标签的查询次数.
 
 ## 2.6.5(2025-01-16)